refactor: reorganize context menu commands and remove unused files
