+++ src/contextMenus/index.ts
@@ -1,6 +1,6 @@
-import * as quote from "./general/quote"
+import * as quote from "./message/quote"
 
-import * as profilePicture from "./general/profilePicture"
+import * as profilePicture from "./user/profilePicture"
 
 const contextMessageMenus = {
     "Créer un citation": quote
